move functions to suitable files
